trying to decouple the theanograf to an independet module. A pain in tha ass
