+++ GeMpy/DataManagement.py
@@ -5,7 +5,7 @@ import theano.tensor as T
 import numpy as np
 import sys, os
 import pandas as pn
-
+import theanograf as tg
 from Visualization import PlotData
 
 
@@ -291,7 +291,8 @@ class DataManagement(object):
             verbose(int): Level of verbosity during the execution of the functions (up to 5). Default 0
         """
 
-        def __init__(self, _data, _grid=None, *args, **kwargs):
+        def __init__(self, _data, _grid=None, compute_block_model=True,
+                     compute_potential_field=False, *args, **kwargs):
 
             verbose = kwargs.get('verbose', 0)
 
@@ -308,132 +309,18 @@ class DataManagement(object):
 
             self._set_constant_parameteres(_data, _grid, **kwargs)
 
-            self.theano_compilation_3D()
-            # -------------------------
-            self.potential_fields = [self.compute_potential_field(i, verbose=verbose)
-                                     for i in np.arange(len(self._data.series.columns))]
+            if compute_potential_field:
+                self.potential_fields = []
+                self._interpolate = self.compile_potential_field_function()
 
-        def update_potential_fields(self, verbose=0):
-            self.potential_fields = [self.compute_potential_field(i, verbose=verbose)
-                                     for i in np.arange(len(self._data.series.columns))]
+                # -------------------------
+                self.potential_fields = [self.compute_potential_fields(i, verbose=verbose)
+                                         for i in np.arange(len(self._data.series.columns))]
 
-        def _select_serie(self, series_name=0):
-            """
-            Return the formations of a given serie in string
-            :param series_name: name or argument of the serie. Default first of the list
-            :return: formations of a given serie in string separeted by |
-            """
-            if type(series_name) == int or type(series_name) == np.int64:
-                _formations_in_serie = "|".join(self._data.series.ix[:, series_name].drop_duplicates())
-            elif type(series_name) == str:
-                _formations_in_serie = "|".join(self._data.series[series_name].drop_duplicates())
-            return _formations_in_serie
+            if compute_block_model:
 
-        def _set_constant_parameteres(self, _data, _grid, **kwargs):
-            """
-            Basic interpolator parameters. Also here it is possible to change some flags of theano
-            :param range_var: Range of the variogram, it is recommended the distance of the longest diagonal
-            :param c_o: Sill of the variogram
-            """
-            # TODO: update Docstrig
-
-            u_grade = kwargs.get('u_grade', 2)
-            range_var = kwargs.get('range_var', None)
-            c_o = kwargs.get('c_o', None)
-            nugget_effect = kwargs.get('nugget_effect', 0.01)
-            rescaling_factor = kwargs.get('rescaling_factor', None)
-
-            if not range_var:
-                range_var = np.sqrt((_data.xmax - _data.xmin) ** 2 +
-                                    (_data.ymax - _data.ymin) ** 2 +
-                                    (_data.zmax - _data.zmin) ** 2)
-            if not c_o:
-                c_o = range_var ** 2 / 14 / 3
-
-            # Creation of shared variables
-            self.nx_T = theano.shared(_data.nx, "Resolution in x axis")
-            self.ny_T = theano.shared(_data.ny, "Resolution in y axis")
-            self.nz_T = theano.shared(_data.nz, "Resolution in z axis")
-            self.a_T = theano.shared(range_var, "range", allow_downcast=True)
-            self.c_o_T = theano.shared(c_o, "covariance at 0", allow_downcast=True)
-            self.nugget_effect_grad_T = theano.shared(nugget_effect, "nugget effect of the grade", allow_downcast=True)
-
-            assert (0 <= u_grade <= 2)
-
-            if u_grade == 0:
-                self.u_grade_T = theano.shared(u_grade, "grade of the universal drift", allow_downcast=True)
-            else:
-                self.u_grade_T = theano.shared(3 ** u_grade, allow_downcast=True)
-            # TODO: To be sure what is the mathematical meaning of this
-
-            if not rescaling_factor:
-                max_coord = pn.concat([_data.foliations, _data.interfaces]).max()[['X', 'Y', 'Z']]
-                min_coord = pn.concat([_data.foliations, _data.interfaces]).min()[['X', 'Y', 'Z']]
-                rescaling_factor = np.max(max_coord - min_coord)
-
-            self.rescaling_factor_T = theano.shared(rescaling_factor, "rescaling factor", allow_downcast=True)
-            self.number_of_points_per_formation_T = theano.shared(np.zeros(2),
-                                                                  "Vector. Number of formations in the serie")
-
-            _universal_matrix = np.vstack((_grid.grid.T,
-                                           (_grid.grid ** 2).T,
-                                           _grid.grid[:, 0] * _grid.grid[:, 1],
-                                           _grid.grid[:, 0] * _grid.grid[:, 2],
-                                           _grid.grid[:, 1] * _grid.grid[:, 2]))
-            self.universal_matrix_T = theano.shared(_universal_matrix + 1e-10, "universal matrix")
-
-            self.block = theano.shared(np.zeros_like(_grid.grid[:, 0]), "Final block computed")
-            self.grid_val_T = theano.shared(_grid.grid + 1e-10, "Positions of the points to interpolate")
-
-        def _get_constant_parameters(self):
-            """
-            Deprecated?
-
-            Returns:
-
-            """
-            return self.a_T, self.c_o_T, self.nugget_effect_grad_T
-
-        def compute_block_model(self, series_number="all", verbose=0):
-            """
-            Method to compute the block model for the given series using data provided in the DataManagement object
-
-            Args:
-                series_number(str or int): series to interpolate
-                verbose(int): level of verbosity during the computation
-
-            Returns:
-                self.block(theano shared[numpy.ndarray]): 3D block with the corresponding formations
-
-            """
-
-            if series_number == "all":
-                series_number = np.arange(len(self._data.series.columns))
-            for i in series_number:
-                formations_in_serie = self._select_serie(i)
-                # Number assigned to each formation
-                n_formation = np.squeeze(np.where(np.in1d(self._data.formations, self._data.series.ix[:, i]))) + 1
-                if verbose > 0:
-                    print(n_formation)
-                self._aux_computations_block_model(formations_in_serie, np.array(n_formation, ndmin=1),
-                                                   verbose=verbose)
-
-        def compute_potential_field(self, series_name=0, verbose=0):
-            """
-            Compute an individual potential field.
-
-            Args:
-                series_name (str or int): name or number of series to interpolate
-                verbose(int): int level of verbosity during the computation
-
-            Returns:
-                numpy.ndarray: 3D array with the potential field
-
-            """
-
-            assert series_name is not "all", "Compute potential field only returns one potential field at the time"
-            formations_in_serie = self._select_serie(series_name)
-            return self._aux_computations_potential_field(formations_in_serie, verbose=verbose)
+                self._block_export = self.compile_block_model_function()
+                self.compute_block_model()
 
         def _aux_computations_block_model(self, for_in_ser, n_formation, verbose=0):
             """
@@ -457,8 +344,8 @@ class DataManagement(object):
                 print("I am in the except")
             # TODO: change [:,:3] that is positional based for XYZ so is more consistent
             dips_position = self._data.foliations[
-                                self._data.foliations["formation"].str.contains(for_in_ser)]\
-                                [['X', 'Y', 'Z']].as_matrix()
+                self._data.foliations["formation"].str.contains(for_in_ser)] \
+                [['X', 'Y', 'Z']].as_matrix()
             dip_angles = self._data.foliations[
                 self._data.foliations["formation"].str.contains(for_in_ser)]["dip"].as_matrix()
             azimuth = self._data.foliations[
@@ -467,9 +354,9 @@ class DataManagement(object):
                 self._data.foliations["formation"].str.contains(for_in_ser)]["polarity"].as_matrix()
 
             if for_in_ser.count("|") == 0:
-                #layers = self._data.interfaces[self._data.interfaces["formation"].str.contains(for_in_ser)].as_matrix()[
+                # layers = self._data.interfaces[self._data.interfaces["formation"].str.contains(for_in_ser)].as_matrix()[
                 #         :, :3]
-                layers = self._data.interfaces[self._data.interfaces["formation"] == for_in_ser]\
+                layers = self._data.interfaces[self._data.interfaces["formation"] == for_in_ser] \
                     [['X', 'Y', 'Z']].as_matrix()
                 rest_layer_points = layers[1:]
                 # TODO self.n_formation probably should not be self
@@ -507,9 +394,9 @@ class DataManagement(object):
                     print('number_formations', n_formation)
                     print('rest_layer_points', rest_layer_points)
 
-            self.grad = self._block_export(dips_position, dip_angles, azimuth, polarity,
-                                           rest_layer_points, ref_layer_points,
-                                           n_formation, yet_simulated)
+            return self._block_export(dips_position, dip_angles, azimuth, polarity,
+                                      rest_layer_points, ref_layer_points,
+                                      n_formation, yet_simulated)
 
         def _aux_computations_potential_field(self, for_in_ser, verbose=0):
             """
@@ -525,8 +412,8 @@ class DataManagement(object):
 
             # TODO: change [:,:3] that is positional based for XYZ so is more consistent
             dips_position = self._data.foliations[
-                                self._data.foliations["formation"].str.contains(for_in_ser)]\
-                                [['X', 'Y', 'Z']].as_matrix()
+                self._data.foliations["formation"].str.contains(for_in_ser)] \
+                [['X', 'Y', 'Z']].as_matrix()
             dip_angles = self._data.foliations[self._data.foliations["formation"].str.contains(for_in_ser)][
                 "dip"].as_matrix()
             azimuth = self._data.foliations[self._data.foliations["formation"].str.contains(for_in_ser)][
@@ -535,7 +422,7 @@ class DataManagement(object):
                 "polarity"].as_matrix()
 
             if for_in_ser.count("|") == 0:
-                #layers = self._data.interfaces[self._data.interfaces["formation"].str.contains(for_in_ser)].as_matrix()[
+                # layers = self._data.interfaces[self._data.interfaces["formation"].str.contains(for_in_ser)].as_matrix()[
                 #         :, :3]
                 layers = self._data.interfaces[self._data.interfaces["formation"] == for_in_ser] \
                     [['X', 'Y', 'Z']].as_matrix()
@@ -562,7 +449,7 @@ class DataManagement(object):
 
             self.Z_x, G_x, G_y, G_z, self.potential_interfaces, self.C, self.DK = self._interpolate(
                 dips_position, dip_angles, azimuth, polarity,
-                rest_layer_points, ref_layer_points)[:]
+                rest_layer_points, ref_layer_points, self.u_grade)[:]
 
             potential_field = self.Z_x.reshape(self._data.nx, self._data.ny, self._data.nz)
 
@@ -574,509 +461,670 @@ class DataManagement(object):
 
             return potential_field
 
-        def theano_compilation_3D(self):
+        def _select_serie(self, series_name=0):
             """
-            Function that generates the symbolic code to perform the interpolation. Calling this function creates
-             both the theano functions for the potential field and the block.
+            Return the formations of a given serie in string
+            :param series_name: name or argument of the serie. Default first of the list
+            :return: formations of a given serie in string separeted by |
+            """
+            if type(series_name) == int or type(series_name) == np.int64:
+                _formations_in_serie = "|".join(self._data.series.ix[:, series_name].drop_duplicates())
+            elif type(series_name) == str:
+                _formations_in_serie = "|".join(self._data.series[series_name].drop_duplicates())
+            return _formations_in_serie
+
+        def _set_constant_parameteres(self, _data, _grid, **kwargs):
+            """
+            Basic interpolator parameters. Also here it is possible to change some flags of theano
+            :param range_var: Range of the variogram, it is recommended the distance of the longest diagonal
+            :param c_o: Sill of the variogram
+            """
+            # TODO: update Docstrig
+
+            u_grade = kwargs.get('u_grade', 2)
+            range_var = kwargs.get('range_var', None)
+            c_o = kwargs.get('c_o', None)
+            nugget_effect = kwargs.get('nugget_effect', 0.01)
+            rescaling_factor = kwargs.get('rescaling_factor', None)
+            self.u_grade = u_grade
+            if not range_var:
+                range_var = np.sqrt((_data.xmax - _data.xmin) ** 2 +
+                                    (_data.ymax - _data.ymin) ** 2 +
+                                    (_data.zmax - _data.zmin) ** 2)
+            if not c_o:
+                c_o = range_var ** 2 / 14 / 3
+
+            # Creation of shared variables
+          #  tg.nx_T = theano.shared(_data.nx, "Resolution in x axis")
+          #  tg.ny_T = theano.shared(_data.ny, "Resolution in y axis")
+          #  tg.nz_T = theano.shared(_data.nz, "Resolution in z axis")
+           # tg.a_T = theano.shared(range_var, "range", allow_downcast=True)
+           # tg.c_o_T = theano.shared(c_o, "covariance at 0", allow_downcast=True)
+           # tg.nugget_effect_grad_T = theano.shared(nugget_effect, "nugget effect of the grade", allow_downcast=True)
+            tg.a_T.set_value(range_var)
+            tg.c_o_T.set_value(c_o)
+            tg.nugget_effect_grad_T.set_value(nugget_effect)
+            assert (0 <= u_grade <= 2)
+
+            if u_grade == 0:
+                print('I am here')
+                tg.u_grade_T.set_value(u_grade)# = theano.shared(u_grade, "grade of the universal drift", allow_downcast=True)
+            else:
+                tg.u_grade_T.set_value(3**u_grade)# = theano.shared(3 ** u_grade, allow_downcast=True)
+            # TODO: To be sure what is the mathematical meaning of this
+
+            if not rescaling_factor:
+                max_coord = pn.concat([_data.foliations, _data.interfaces]).max()[['X', 'Y', 'Z']]
+                min_coord = pn.concat([_data.foliations, _data.interfaces]).min()[['X', 'Y', 'Z']]
+                rescaling_factor = np.max(max_coord - min_coord)
+
+            tg.rescaling_factor_T.set_value(rescaling_factor)# = theano.shared(rescaling_factor, "rescaling factor", allow_downcast=True)
+          #  tg.number_of_points_per_formation_T.set_value = theano.shared(np.zeros(2),
+           #                                                       "Vector. Number of formations in the serie")
+
+            _universal_matrix = np.vstack((_grid.grid.T,
+                                           (_grid.grid ** 2).T,
+                                           _grid.grid[:, 0] * _grid.grid[:, 1],
+                                           _grid.grid[:, 0] * _grid.grid[:, 2],
+                                           _grid.grid[:, 1] * _grid.grid[:, 2]))
+         #   tg.universal_matrix_T = theano.shared(_universal_matrix + 1e-10, "universal matrix")
+            tg.universal_matrix_T.set_value(_universal_matrix + 1e-10)
+            tg.final_block.set_value(np.zeros_like(_grid.grid[:, 0])) # = theano.shared(np.zeros_like(_grid.grid[:, 0]), "Final block computed")
+            # tg.grid_val_T = theano.shared(_grid.grid + 1e-10, "Positions of the points to interpolate")
+            tg.grid_val_T.set_value(_grid.grid + 10e-6)
+
+        def _get_constant_parameters(self):
+            """
+            Deprecated?
 
             Returns:
-                theano function for the potential field
-                theano function for the block
+
+            """
+            return self.a_T, self.c_o_T, self.nugget_effect_grad_T
+
+        def compile_potential_field_function(self):
+            self._interpolate = theano.function(
+                [tg.dips_position, tg.dip_angles, tg.azimuth, tg.polarity, tg.rest_layer_points,
+                 tg.ref_layer_points, tg.u_grade_T, theano.In(tg.yet_simulated, value=np.ones_like(self._grid.grid[:, 0]))],
+                [tg.Z_x, tg.G_x, tg.G_y, tg.G_z, tg.potential_field_interfaces, tg.C_matrix, tg.printing],
+                on_unused_input="warn", profile=True, allow_input_downcast=True)
+            return self._interpolate
+
+        def compile_block_model_function(self):
+
+            self._block_export = theano.function([tg.dips_position, tg.dip_angles, tg.azimuth, tg.polarity,
+                                                  tg.rest_layer_points, tg.ref_layer_points, tg.n_formation,
+                                                  tg.yet_simulated],
+                                                 None,
+                                                 updates=[(tg.final_block, tg.potential_field_contribution)],
+                                                 on_unused_input="warn", profile=True, allow_input_downcast=True,)
+            #   mode=theano.compile.MonitorMode(
+            #   post_func=detect_nan))
+            #    mode=NanGuardMode(nan_is_error=True, inf_is_error=True,
+            #                     big_is_error=True))
+
+            return self._block_export
+
+        def update_potential_fields(self, verbose=0):
+            self.potential_fields = [self.compute_potential_fields(i, verbose=verbose)
+                                     for i in np.arange(len(self._data.series.columns))]
+
+        def compute_block_model(self, series_number="all", verbose=0):
+            """
+            Method to compute the block model for the given series using data provided in the DataManagement object
+
+            Args:
+                series_number(str or int): series to interpolate
+                verbose(int): level of verbosity during the computation
+
+            Returns:
+                self.block(theano shared[numpy.ndarray]): 3D block with the corresponding formations
+
             """
 
-            # Creation of symbolic variables
-            dips_position = T.matrix("Position of the dips")
-            dip_angles = T.vector("Angle of every dip")
-            azimuth = T.vector("Azimuth")
-            polarity = T.vector("Polarity")
-            ref_layer_points = T.matrix("Reference points for every layer")
-            rest_layer_points = T.matrix("Rest of the points of the layers")
+            if series_number == "all":
+                series_number = np.arange(len(self._data.series.columns))
+            for i in series_number:
+                formations_in_serie = self._select_serie(i)
+                # Number assigned to each formation
+                n_formation = np.squeeze(np.where(np.in1d(self._data.formations, self._data.series.ix[:, i]))) + 1
+                if verbose > 0:
+                    print(n_formation)
+                self.grad = self._aux_computations_block_model(formations_in_serie, np.array(n_formation, ndmin=1),
+                                                               verbose=verbose)
 
-            # Init values
-            n_dimensions = 3
-            grade_universal = self.u_grade_T
+                return self.block
 
-            # Calculating the dimensions of the
-            length_of_CG = dips_position.shape[0] * n_dimensions
-            length_of_CGI = rest_layer_points.shape[0]
-            length_of_U_I = grade_universal
-            length_of_C = length_of_CG + length_of_CGI + length_of_U_I
+        def compute_potential_fields(self, series_name="all", verbose=0):
+            """
+            Compute an individual potential field.
 
-            # Extra parameters
-            i_reescale = 1 / (self.rescaling_factor_T ** 2)
-            gi_reescale = 1 / self.rescaling_factor_T
+            Args:
+                series_name (str or int): name or number of series to interpolate
+                verbose(int): int level of verbosity during the computation
 
-            # TODO: Check that the distances does not go nuts when I use too large numbers
+            Returns:
+                numpy.ndarray: 3D array with the potential field
 
-            # Here we create the array with the points to simulate:
-            #   grid points except those who have been simulated in a younger serie
-            #   interfaces points to segment the lithologies
-            yet_simulated = T.vector("boolean function that avoid to simulate twice a point of a different serie")
-            grid_val = T.vertical_stack((self.grid_val_T * yet_simulated.reshape(
-                                        (yet_simulated.shape[0], 1))).nonzero_values().reshape((-1, 3)),
-                                        rest_layer_points)
+            """
 
-            # ==========================================
-            # Calculation of Cartesian and Euclidian distances
-            # ===========================================
-            # Auxiliary tile for dips and transformation to float 64 of variables in order to calculate
-            #  precise euclidian
-            # distances
-            _aux_dips_pos = T.tile(dips_position, (n_dimensions, 1)).astype("float64")
-            _aux_rest_layer_points = rest_layer_points.astype("float64")
-            _aux_ref_layer_points = ref_layer_points.astype("float64")
-            _aux_grid_val = grid_val.astype("float64")
-
-            # Calculation of euclidian distances giving back float32
-            SED_rest_rest = (T.sqrt(
-                (_aux_rest_layer_points ** 2).sum(1).reshape((_aux_rest_layer_points.shape[0], 1)) +
-                (_aux_rest_layer_points ** 2).sum(1).reshape((1, _aux_rest_layer_points.shape[0])) -
-                2 * _aux_rest_layer_points.dot(_aux_rest_layer_points.T))).astype("float32")
-
-            SED_ref_rest = (T.sqrt(
-                (_aux_ref_layer_points ** 2).sum(1).reshape((_aux_ref_layer_points.shape[0], 1)) +
-                (_aux_rest_layer_points ** 2).sum(1).reshape((1, _aux_rest_layer_points.shape[0])) -
-                2 * _aux_ref_layer_points.dot(_aux_rest_layer_points.T))).astype("float32")
-
-            SED_rest_ref = (T.sqrt(
-                (_aux_rest_layer_points ** 2).sum(1).reshape((_aux_rest_layer_points.shape[0], 1)) +
-                (_aux_ref_layer_points ** 2).sum(1).reshape((1, _aux_ref_layer_points.shape[0])) -
-                2 * _aux_rest_layer_points.dot(_aux_ref_layer_points.T))).astype("float32")
-
-            SED_ref_ref = (T.sqrt(
-                (_aux_ref_layer_points ** 2).sum(1).reshape((_aux_ref_layer_points.shape[0], 1)) +
-                (_aux_ref_layer_points ** 2).sum(1).reshape((1, _aux_ref_layer_points.shape[0])) -
-                2 * _aux_ref_layer_points.dot(_aux_ref_layer_points.T))).astype("float32")
-
-            SED_dips_dips = (T.sqrt(
-                (_aux_dips_pos ** 2).sum(1).reshape((_aux_dips_pos.shape[0], 1)) +
-                (_aux_dips_pos ** 2).sum(1).reshape((1, _aux_dips_pos.shape[0])) -
-                2 * _aux_dips_pos.dot(_aux_dips_pos.T))).astype("float32")
-
-            SED_dips_rest = (T.sqrt(
-                (_aux_dips_pos ** 2).sum(1).reshape((_aux_dips_pos.shape[0], 1)) +
-                (_aux_rest_layer_points ** 2).sum(1).reshape((1, _aux_rest_layer_points.shape[0])) -
-                2 * _aux_dips_pos.dot(_aux_rest_layer_points.T))).astype("float32")
-
-            SED_dips_ref = (T.sqrt(
-                (_aux_dips_pos ** 2).sum(1).reshape((_aux_dips_pos.shape[0], 1)) +
-                (_aux_ref_layer_points ** 2).sum(1).reshape((1, _aux_ref_layer_points.shape[0])) -
-                2 * _aux_dips_pos.dot(_aux_ref_layer_points.T))).astype("float32")
-
-            # Calculating euclidian distances between the point to simulate and the avalible data
-            SED_dips_SimPoint = (T.sqrt(
-                (_aux_dips_pos ** 2).sum(1).reshape((_aux_dips_pos.shape[0], 1)) +
-                (_aux_grid_val ** 2).sum(1).reshape((1, _aux_grid_val.shape[0])) -
-                2 * _aux_dips_pos.dot(_aux_grid_val.T))).astype("float32")
-
-            SED_rest_SimPoint = (T.sqrt(
-                (_aux_rest_layer_points ** 2).sum(1).reshape((_aux_rest_layer_points.shape[0], 1)) +
-                (_aux_grid_val ** 2).sum(1).reshape((1, _aux_grid_val.shape[0])) -
-                2 * _aux_rest_layer_points.dot(_aux_grid_val.T))).astype("float32")
-
-            SED_ref_SimPoint = (T.sqrt(
-                (_aux_ref_layer_points ** 2).sum(1).reshape((_aux_ref_layer_points.shape[0], 1)) +
-                (_aux_grid_val ** 2).sum(1).reshape((1, _aux_grid_val.shape[0])) -
-                2 * _aux_ref_layer_points.dot(_aux_grid_val.T))).astype("float32")
-
-            # Cartesian distances between dips positions
-            h_u = T.vertical_stack(
-                T.tile(dips_position[:, 0] - dips_position[:, 0].reshape((dips_position[:, 0].shape[0], 1)),
-                       n_dimensions),
-                T.tile(dips_position[:, 1] - dips_position[:, 1].reshape((dips_position[:, 1].shape[0], 1)),
-                       n_dimensions),
-                T.tile(dips_position[:, 2] - dips_position[:, 2].reshape((dips_position[:, 2].shape[0], 1)),
-                       n_dimensions))
-
-            h_v = h_u.T
-
-            # Cartesian distances between dips and interface points
-            # Rest
-            hu_rest = T.vertical_stack(
-                (dips_position[:, 0] - rest_layer_points[:, 0].reshape((rest_layer_points[:, 0].shape[0], 1))).T,
-                (dips_position[:, 1] - rest_layer_points[:, 1].reshape((rest_layer_points[:, 1].shape[0], 1))).T,
-                (dips_position[:, 2] - rest_layer_points[:, 2].reshape((rest_layer_points[:, 2].shape[0], 1))).T
-            )
-
-            # Reference point
-            hu_ref = T.vertical_stack(
-                (dips_position[:, 0] - ref_layer_points[:, 0].reshape((ref_layer_points[:, 0].shape[0], 1))).T,
-                (dips_position[:, 1] - ref_layer_points[:, 1].reshape((ref_layer_points[:, 1].shape[0], 1))).T,
-                (dips_position[:, 2] - ref_layer_points[:, 2].reshape((ref_layer_points[:, 2].shape[0], 1))).T
-            )
-
-            # Cartesian distances between reference points and rest
-            hx = T.stack(
-                (rest_layer_points[:, 0] - ref_layer_points[:, 0]),
-                (rest_layer_points[:, 1] - ref_layer_points[:, 1]),
-                (rest_layer_points[:, 2] - ref_layer_points[:, 2])
-            ).T
-
-            # Cartesian distances between the point to simulate and the dips
-            hu_SimPoint = T.vertical_stack(
-                (dips_position[:, 0] - grid_val[:, 0].reshape((grid_val[:, 0].shape[0], 1))).T,
-                (dips_position[:, 1] - grid_val[:, 1].reshape((grid_val[:, 1].shape[0], 1))).T,
-                (dips_position[:, 2] - grid_val[:, 2].reshape((grid_val[:, 2].shape[0], 1))).T
-            )
-
-            # Perpendicularity matrix. Boolean matrix to separate cross-covariance and
-            # every gradient direction covariance
-            perpendicularity_matrix = T.zeros_like(SED_dips_dips)
-
-            # Cross-covariances of x
-            perpendicularity_matrix = T.set_subtensor(
-                perpendicularity_matrix[0:dips_position.shape[0], 0:dips_position.shape[0]], 1)
-
-            # Cross-covariances of y
-            perpendicularity_matrix = T.set_subtensor(
-                perpendicularity_matrix[dips_position.shape[0]:dips_position.shape[0] * 2,
-                dips_position.shape[0]:dips_position.shape[0] * 2], 1)
-
-            # Cross-covariances of y
-            perpendicularity_matrix = T.set_subtensor(
-                perpendicularity_matrix[dips_position.shape[0] * 2:dips_position.shape[0] * 3,
-                dips_position.shape[0] * 2:dips_position.shape[0] * 3], 1)
-
-
-           # printing = (self.c_o_T * i_reescale * (
-           #     (SED_rest_rest < self.a_T) *  # Rest - Rest Covariances Matrix
-            #printing = (1 - 7 * (SED_rest_rest / self.a_T) ** 2 +
-            #     35 / 4 * (SED_rest_rest / self.a_T) ** 3 -
-            #     7 / 2 * (SED_rest_rest / self.a_T) ** 5 +
-            #     3 / 4 * (SED_rest_rest / self.a_T) ** 7)
-            printing = SED_rest_rest*self.a_T
-            # ==========================
-            # Creating covariance Matrix
-            # ==========================
-            # Covariance matrix for interfaces
-            C_I = (self.c_o_T * i_reescale * (
-                (SED_rest_rest < self.a_T) *  # Rest - Rest Covariances Matrix
-                (1 - 7 * (SED_rest_rest / self.a_T) ** 2 +
-                 35 / 4 * (SED_rest_rest / self.a_T) ** 3 -
-                 7 / 2 * (SED_rest_rest / self.a_T) ** 5 +
-                 3 / 4 * (SED_rest_rest / self.a_T) ** 7) -
-                ((SED_ref_rest < self.a_T) *  # Reference - Rest
-                 (1 - 7 * (SED_ref_rest / self.a_T) ** 2 +
-                  35 / 4 * (SED_ref_rest / self.a_T) ** 3 -
-                  7 / 2 * (SED_ref_rest / self.a_T) ** 5 +
-                  3 / 4 * (SED_ref_rest / self.a_T) ** 7)) -
-                ((SED_rest_ref < self.a_T) *  # Rest - Reference
-                 (1 - 7 * (SED_rest_ref / self.a_T) ** 2 +
-                  35 / 4 * (SED_rest_ref / self.a_T) ** 3 -
-                  7 / 2 * (SED_rest_ref / self.a_T) ** 5 +
-                  3 / 4 * (SED_rest_ref / self.a_T) ** 7)) +
-                ((SED_ref_ref < self.a_T) *  # Reference - References
-                 (1 - 7 * (SED_ref_ref / self.a_T) ** 2 +
-                  35 / 4 * (SED_ref_ref / self.a_T) ** 3 -
-                  7 / 2 * (SED_ref_ref / self.a_T) ** 5 +
-                  3 / 4 * (SED_ref_ref / self.a_T) ** 7)))) #'+ 10e-6
-
-            SED_dips_dips = T.switch(T.eq(SED_dips_dips, 0), 1, SED_dips_dips)
-
-            # Covariance matrix for gradients at every xyz direction and their cross-covariances
-            C_G = T.switch(
-                T.eq(SED_dips_dips, 0),  # This is the condition
-                0,  # If true it is equal to 0. This is how a direction affect another
-                (  # else, following Chiles book
-                    (h_u * h_v / SED_dips_dips ** 2) *
-                    ((
-                         (SED_dips_dips < self.a_T) *  # first derivative
-                         (-self.c_o_T * ((-14 / self.a_T ** 2) + 105 / 4 * SED_dips_dips / self.a_T ** 3 -
-                                         35 / 2 * SED_dips_dips ** 3 / self.a_T ** 5 +
-                                         21 / 4 * SED_dips_dips ** 5 / self.a_T ** 7))) +
-                     (SED_dips_dips < self.a_T) *  # Second derivative
-                     self.c_o_T * 7 * (9 * SED_dips_dips ** 5 - 20 * self.a_T ** 2 * SED_dips_dips ** 3 +
-                                       15 * self.a_T ** 4 * SED_dips_dips - 4 * self.a_T ** 5) / (2 * self.a_T ** 7)) -
-                    (perpendicularity_matrix *
-                     (SED_dips_dips < self.a_T) *  # first derivative
-                     self.c_o_T * ((-14 / self.a_T ** 2) + 105 / 4 * SED_dips_dips / self.a_T ** 3 -
-                                   35 / 2 * SED_dips_dips ** 3 / self.a_T ** 5 +
-                                   21 / 4 * SED_dips_dips ** 5 / self.a_T ** 7)))
-            )
-
-            # Setting nugget effect of the gradients
-            # TODO: This function can be substitued by simply adding the nugget effect to the diag
-            C_G = T.fill_diagonal(C_G, -self.c_o_T * (-14 / self.a_T ** 2) + self.nugget_effect_grad_T)
-
-            # Cross-Covariance gradients-interfaces
-            C_GI = gi_reescale * (
-                (hu_rest *
-                 (SED_dips_rest < self.a_T) *  # first derivative
-                 (- self.c_o_T * ((-14 / self.a_T ** 2) + 105 / 4 * SED_dips_rest / self.a_T ** 3 -
-                                  35 / 2 * SED_dips_rest ** 3 / self.a_T ** 5 +
-                                  21 / 4 * SED_dips_rest ** 5 / self.a_T ** 7))) -
-                (hu_ref *
-                 (SED_dips_ref < self.a_T) *  # first derivative
-                 (- self.c_o_T * ((-14 / self.a_T ** 2) + 105 / 4 * SED_dips_ref / self.a_T ** 3 -
-                                  35 / 2 * SED_dips_ref ** 3 / self.a_T ** 5 +
-                                  21 / 4 * SED_dips_ref ** 5 / self.a_T ** 7)))
-            ).T
-
-            if self.u_grade_T.get_value() == 3:
-                # ==========================
-                # Condition of universality 1 degree
-
-                # Gradients
-                n = dips_position.shape[0]
-                U_G = T.zeros((n * n_dimensions, n_dimensions))
-                # x
-                U_G = T.set_subtensor(
-                    U_G[:n, 0], 1)
-                # y
-                U_G = T.set_subtensor(
-                    U_G[n:n * 2, 1], 1
-                )
-                # z
-                U_G = T.set_subtensor(
-                    U_G[n * 2: n * 3, 2], 1
-                )
-
-            #    U_G = T.set_subtensor(U_G[:, -1], [0, 0, 0, 0, 0, 0])
-
-                # Interface
-                U_I = -hx * gi_reescale
-
-            #    hxf = (T.lt(rest_layer_points[:, 0], 5) - T.lt(ref_layer_points[:, 0], 5))*2 + 1
-
-            #    U_I = T.horizontal_stack(U_I, T.stack(hxf).T)
-
-            elif self.u_grade_T.get_value() == 9:
-                # ==========================
-                # Condition of universality 2 degree
-                # Gradients
-
-                n = dips_position.shape[0]
-                U_G = T.zeros((n * n_dimensions, 3 * n_dimensions))
-                # x
-                U_G = T.set_subtensor(U_G[:n, 0], 1)
-                # y
-                U_G = T.set_subtensor(U_G[n * 1:n * 2, 1], 1)
-                # z
-                U_G = T.set_subtensor(U_G[n * 2: n * 3, 2], 1)
-                # x**2
-                U_G = T.set_subtensor(U_G[:n, 3], 2 * gi_reescale * dips_position[:, 0])
-                # y**2
-                U_G = T.set_subtensor(U_G[n * 1:n * 2, 4], 2 * gi_reescale * dips_position[:, 1])
-                # z**2
-                U_G = T.set_subtensor(U_G[n * 2: n * 3, 5], 2 * gi_reescale * dips_position[:, 2])
-                # xy
-                U_G = T.set_subtensor(U_G[:n, 6], gi_reescale * dips_position[:, 1])  # This is y
-                U_G = T.set_subtensor(U_G[n * 1:n * 2, 6], gi_reescale * dips_position[:, 0])  # This is x
-                # xz
-                U_G = T.set_subtensor(U_G[:n, 7], gi_reescale * dips_position[:, 2])  # This is z
-                U_G = T.set_subtensor(U_G[n * 2: n * 3, 7], gi_reescale * dips_position[:, 0])  # This is x
-                # yz
-                U_G = T.set_subtensor(U_G[n * 1:n * 2, 8], gi_reescale * dips_position[:, 2])  # This is z
-                U_G = T.set_subtensor(U_G[n * 2:n * 3, 8], gi_reescale * dips_position[:, 1])  # This is y
-
-                # Interface
-                U_I = - T.stack(
-                    gi_reescale * (rest_layer_points[:, 0] - ref_layer_points[:, 0]), # x
-                    gi_reescale * (rest_layer_points[:, 1] - ref_layer_points[:, 1]), # y
-                    gi_reescale * (rest_layer_points[:, 2] - ref_layer_points[:, 2]), # z
-                    gi_reescale ** 2 * (rest_layer_points[:, 0] ** 2 - ref_layer_points[:, 0] ** 2), # xx
-                    gi_reescale ** 2 * (rest_layer_points[:, 1] ** 2 - ref_layer_points[:, 1] ** 2), # yy
-                    gi_reescale ** 2 * (rest_layer_points[:, 2] ** 2 - ref_layer_points[:, 2] ** 2), # zz
-                    gi_reescale ** 2 * (rest_layer_points[:, 0] * rest_layer_points[:, 1] - ref_layer_points[:, 0] *
-                                        ref_layer_points[:, 1]),
-                    gi_reescale ** 2 * (rest_layer_points[:, 0] * rest_layer_points[:, 2] - ref_layer_points[:, 0] *
-                                        ref_layer_points[:, 2]),
-                    gi_reescale ** 2 * (rest_layer_points[:, 1] * rest_layer_points[:, 2] - ref_layer_points[:, 1] *
-                                        ref_layer_points[:, 2]),
-                ).T
-
-
-
-            # =================================
-            # Creation of the Covariance Matrix
-            # =================================
-            C_matrix = T.zeros((length_of_C, length_of_C ))
-
-            # First row of matrices
-            C_matrix = T.set_subtensor(C_matrix[0:length_of_CG, 0:length_of_CG], C_G)
-
-            C_matrix = T.set_subtensor(C_matrix[0:length_of_CG, length_of_CG:length_of_CG + length_of_CGI], C_GI.T)
-
-            if not self.u_grade_T.get_value() == 0:
-                C_matrix = T.set_subtensor(C_matrix[0:length_of_CG, -length_of_U_I:], U_G)
-
-            # Second row of matrices
-            C_matrix = T.set_subtensor(C_matrix[length_of_CG:length_of_CG + length_of_CGI, 0:length_of_CG], C_GI)
-            C_matrix = T.set_subtensor(C_matrix[length_of_CG:length_of_CG + length_of_CGI,
-                                       length_of_CG:length_of_CG + length_of_CGI], C_I)
-
-            if not self.u_grade_T.get_value() == 0:
-                C_matrix = T.set_subtensor(C_matrix[length_of_CG:length_of_CG + length_of_CGI, -length_of_U_I:], U_I)
-
-                # Third row of matrices
-                C_matrix = T.set_subtensor(C_matrix[-length_of_U_I:, 0:length_of_CG], U_G.T)
-                C_matrix = T.set_subtensor(C_matrix[-length_of_U_I:, length_of_CG:length_of_CG + length_of_CGI], U_I.T)
-
-            # =====================
-            # Creation of the gradients G vector
-            # Calculation of the cartesian components of the dips assuming the unit module
-            G_x = T.sin(T.deg2rad(dip_angles)) * T.sin(T.deg2rad(azimuth)) * polarity
-            G_y = T.sin(T.deg2rad(dip_angles)) * T.cos(T.deg2rad(azimuth)) * polarity
-            G_z = T.cos(T.deg2rad(dip_angles)) * polarity
-
-            G = T.concatenate((G_x, G_y, G_z))
-
-            # Creation of the Dual Kriging vector
-            b = T.zeros_like(C_matrix[:, 0])
-            b = T.set_subtensor(b[0:G.shape[0]], G)
-
-            # Solving the kriging system
-            # TODO: look for an eficient way to substitute nlianlg by a theano operation
-            DK_parameters = T.dot(T.nlinalg.matrix_inverse(C_matrix), b)
-
-            # ==============
-            # Interpolator
-            # ==============
-
-            # Creation of a matrix of dimensions equal to the grid with the weights for every point (big 4D matrix in
-            # ravel form)
-            weights = T.tile(DK_parameters, (grid_val.shape[0], 1)).T
-
-            # Gradient contribution
-            sigma_0_grad = T.sum(
-                (weights[:length_of_CG, :] *
-                 gi_reescale *
-                 (-hu_SimPoint *
-                  (SED_dips_SimPoint < self.a_T) *  # first derivative
-                  (- self.c_o_T * ((-14 / self.a_T ** 2) + 105 / 4 * SED_dips_SimPoint / self.a_T ** 3 -
-                                   35 / 2 * SED_dips_SimPoint ** 3 / self.a_T ** 5 +
-                                   21 / 4 * SED_dips_SimPoint ** 5 / self.a_T ** 7)))),
-                axis=0)
-
-            # Interface contribution
-            sigma_0_interf = (T.sum(
-                -weights[length_of_CG:length_of_CG + length_of_CGI, :] *
-                (self.c_o_T * i_reescale * (
-                    (SED_rest_SimPoint < self.a_T) *  # SimPoint - Rest Covariances Matrix
-                    (1 - 7 * (SED_rest_SimPoint / self.a_T) ** 2 +
-                     35 / 4 * (SED_rest_SimPoint / self.a_T) ** 3 -
-                     7 / 2 * (SED_rest_SimPoint / self.a_T) ** 5 +
-                     3 / 4 * (SED_rest_SimPoint / self.a_T) ** 7) -
-                    ((SED_ref_SimPoint < self.a_T) *  # SimPoint- Ref
-                     (1 - 7 * (SED_ref_SimPoint / self.a_T) ** 2 +
-                      35 / 4 * (SED_ref_SimPoint / self.a_T) ** 3 -
-                      7 / 2 * (SED_ref_SimPoint / self.a_T) ** 5 +
-                      3 / 4 * (SED_ref_SimPoint / self.a_T) ** 7)))), axis=0))
-
-            # Universal drift contribution
-            # Universal terms used to calculate f0
-            _universal_terms_layers = T.horizontal_stack(
-                rest_layer_points,
-                (rest_layer_points ** 2),
-                T.stack((rest_layer_points[:, 0] * rest_layer_points[:, 1],
-                         rest_layer_points[:, 0] * rest_layer_points[:, 2],
-                         rest_layer_points[:, 1] * rest_layer_points[:, 2]), axis=1)).T
-
-            universal_matrix = T.horizontal_stack(
-                (self.universal_matrix_T * yet_simulated).nonzero_values().reshape((9, -1)),
-                _universal_terms_layers)
-
-            if self.u_grade_T.get_value() == 0:
-                f_0 = 0
+            #assert series_name is not "all", "Compute potential field only returns one potential field at the time"
+            if series_name is 'all':
+                for i in np.arange(len(self._data.series.columns)):
+                    formations_in_serie = self._select_serie(i)
+                    self.potential_fields.append(self._aux_computations_potential_field(formations_in_serie,
+                                                                                        verbose=verbose))
             else:
-                gi_rescale_aux = T.repeat(gi_reescale, 9)
-                gi_rescale_aux = T.set_subtensor(gi_rescale_aux[:3], 1)
-                _aux_magic_term = T.tile(gi_rescale_aux[:grade_universal], (grid_val.shape[0], 1)).T
-                f_0 = (T.sum(
-                    weights[-length_of_U_I:, :] * gi_reescale * _aux_magic_term *
-                    universal_matrix[:grade_universal]
-                    , axis=0))
+                formations_in_serie = self._select_serie(series_name)
+                self.potential_fields = self._aux_computations_potential_field(formations_in_serie, verbose=verbose)
 
-            # Contribution faults
-           # f_1 = weights[-1, :] * T.lt(universal_matrix[0, :], 5) * 2 - 1
-         #   f_1 = 0
-            # Potential field
-            # Value of the potential field
+            return self.potential_fields
 
-            Z_x = (sigma_0_grad + sigma_0_interf + f_0)[:-rest_layer_points.shape[0]]
-            potential_field_interfaces = (sigma_0_grad + sigma_0_interf + f_0)[-rest_layer_points.shape[0]:]
 
-            # Theano function to calculate a potential field
-            self._interpolate = theano.function(
-                [dips_position, dip_angles, azimuth, polarity, rest_layer_points, ref_layer_points,
-                 theano.In(yet_simulated, value=np.ones_like(self._grid.grid[:, 0]))],
-                [Z_x, G_x, G_y, G_z, potential_field_interfaces, C_matrix, printing],
-                on_unused_input="warn", profile=True, allow_input_downcast=True)
 
-            # =======================================================================
-            #               CODE TO EXPORT THE BLOCK DIRECTLY
-            # ========================================================================
-
-            # Aux shared parameters
-           # infinite_pos = theano.shared(np.float32(np.inf))
-           # infinite_neg = theano.shared(np.float32(-np.inf))
-
-            # Value of the lithology-segment
-            n_formation = T.vector("The assigned number of the lithologies in this serie")
-
-            # Loop to obtain the average Zx for every intertace
-            def average_potential(dim_a, dim_b, pfi):
-                """
-
-                :param dim: size of the rest values vector per formation
-                :param pfi: the values of all the rest values potentials
-                :return: average of the potential per formation
-                """
-                average = pfi[T.cast(dim_a, "int32"): T.cast(dim_b, "int32")].sum() / (dim_b - dim_a)
-                return average
-
-            potential_field_unique, updates1 = theano.scan(fn=average_potential,
-                                                           outputs_info=None,
-                                                           sequences=dict(
-                                                               input=T.concatenate(
-                                                                   (T.stack(0),
-                                                                    self.number_of_points_per_formation_T)),
-                                                               taps=[0, 1]),
-                                                           non_sequences=potential_field_interfaces)
-
-            infinite_pos = T.max(potential_field_unique) + 10
-            infinite_neg = T.min(potential_field_unique) - 10
-
-            # Loop to segment the distinct lithologies
-            potential_field_iter = T.concatenate((T.stack(infinite_pos),
-                                                  potential_field_unique,
-                                                  T.stack(infinite_neg)))
-
-            def compare(a, b, n_formation, Zx):
-                return T.le(Zx, a) * T.ge(Zx, b) * n_formation
-
-            block, updates2 = theano.scan(fn=compare,
-                                          outputs_info=None,
-                                          sequences=[dict(input=potential_field_iter, taps=[0, 1]),
-                                                     n_formation],
-                                          non_sequences=Z_x)
-
-            # Adding to the block the contribution of the potential field
-            potential_field_contribution = T.set_subtensor(
-                self.block[T.nonzero(T.cast(yet_simulated, "int8"))[0]],
-                block.sum(axis=0))
-
-            # Some gradient testing
-            # grad = T.jacobian(T.flatten(printing), rest_layer_points)
-            grad = T.grad(T.sum(Z_x), self.a_T)
-            from theano.compile.nanguardmode import NanGuardMode
-
-            def detect_nan(i, node, fn):
-                for output in fn.outputs:
-                    if (not isinstance(output[0], np.random.RandomState) and
-                            np.isnan(output[0]).any()):
-                        print('*** NaN detected ***')
-                        theano.printing.debugprint(node)
-                        print('Inputs : %s' % [input[0] for input in fn.inputs])
-                        print('Outputs: %s' % [output[0] for output in fn.outputs])
-                        break
-
-            # Theano function to update the block
-            self._block_export = theano.function([dips_position, dip_angles, azimuth, polarity, rest_layer_points,
-                                                  ref_layer_points, n_formation, yet_simulated], grad,
-                                                 updates=[(self.block, potential_field_contribution)],
-                                                 on_unused_input="warn", profile=True, allow_input_downcast=True,)
-                                             #   mode=theano.compile.MonitorMode(
-                                             #       post_func=detect_nan))
-                                              #    mode=NanGuardMode(nan_is_error=True, inf_is_error=True,
-                                              #                     big_is_error=True))
+        # def theano_compilation_3D(self):
+        #     """
+        #     Function that generates the symbolic code to perform the interpolation. Calling this function creates
+        #      both the theano functions for the potential field and the block.
+        #
+        #     Returns:
+        #         theano function for the potential field
+        #         theano function for the block
+        #     """
+        #
+        #     # Creation of symbolic variables
+        #     dips_position = T.matrix("Position of the dips")
+        #     dip_angles = T.vector("Angle of every dip")
+        #     azimuth = T.vector("Azimuth")
+        #     polarity = T.vector("Polarity")
+        #     ref_layer_points = T.matrix("Reference points for every layer")
+        #     rest_layer_points = T.matrix("Rest of the points of the layers")
+        #
+        #     # Init values
+        #     n_dimensions = 3
+        #     grade_universal = self.u_grade_T
+        #
+        #     # Calculating the dimensions of the
+        #     length_of_CG = dips_position.shape[0] * n_dimensions
+        #     length_of_CGI = rest_layer_points.shape[0]
+        #     length_of_U_I = grade_universal
+        #     length_of_C = length_of_CG + length_of_CGI + length_of_U_I
+        #
+        #     # Extra parameters
+        #     i_reescale = 1 / (self.rescaling_factor_T ** 2)
+        #     gi_reescale = 1 / self.rescaling_factor_T
+        #
+        #     # TODO: Check that the distances does not go nuts when I use too large numbers
+        #
+        #     # Here we create the array with the points to simulate:
+        #     #   grid points except those who have been simulated in a younger serie
+        #     #   interfaces points to segment the lithologies
+        #     yet_simulated = T.vector("boolean function that avoid to simulate twice a point of a different serie")
+        #     grid_val = T.vertical_stack((self.grid_val_T * yet_simulated.reshape(
+        #                                 (yet_simulated.shape[0], 1))).nonzero_values().reshape((-1, 3)),
+        #                                 rest_layer_points)
+        #
+        #     # ==========================================
+        #     # Calculation of Cartesian and Euclidian distances
+        #     # ===========================================
+        #     # Auxiliary tile for dips and transformation to float 64 of variables in order to calculate
+        #     #  precise euclidian
+        #     # distances
+        #     _aux_dips_pos = T.tile(dips_position, (n_dimensions, 1)).astype("float64")
+        #     _aux_rest_layer_points = rest_layer_points.astype("float64")
+        #     _aux_ref_layer_points = ref_layer_points.astype("float64")
+        #     _aux_grid_val = grid_val.astype("float64")
+        #
+        #     # Calculation of euclidian distances giving back float32
+        #     SED_rest_rest = (T.sqrt(
+        #         (_aux_rest_layer_points ** 2).sum(1).reshape((_aux_rest_layer_points.shape[0], 1)) +
+        #         (_aux_rest_layer_points ** 2).sum(1).reshape((1, _aux_rest_layer_points.shape[0])) -
+        #         2 * _aux_rest_layer_points.dot(_aux_rest_layer_points.T))).astype("float32")
+        #
+        #     SED_ref_rest = (T.sqrt(
+        #         (_aux_ref_layer_points ** 2).sum(1).reshape((_aux_ref_layer_points.shape[0], 1)) +
+        #         (_aux_rest_layer_points ** 2).sum(1).reshape((1, _aux_rest_layer_points.shape[0])) -
+        #         2 * _aux_ref_layer_points.dot(_aux_rest_layer_points.T))).astype("float32")
+        #
+        #     SED_rest_ref = (T.sqrt(
+        #         (_aux_rest_layer_points ** 2).sum(1).reshape((_aux_rest_layer_points.shape[0], 1)) +
+        #         (_aux_ref_layer_points ** 2).sum(1).reshape((1, _aux_ref_layer_points.shape[0])) -
+        #         2 * _aux_rest_layer_points.dot(_aux_ref_layer_points.T))).astype("float32")
+        #
+        #     SED_ref_ref = (T.sqrt(
+        #         (_aux_ref_layer_points ** 2).sum(1).reshape((_aux_ref_layer_points.shape[0], 1)) +
+        #         (_aux_ref_layer_points ** 2).sum(1).reshape((1, _aux_ref_layer_points.shape[0])) -
+        #         2 * _aux_ref_layer_points.dot(_aux_ref_layer_points.T))).astype("float32")
+        #
+        #     SED_dips_dips = (T.sqrt(
+        #         (_aux_dips_pos ** 2).sum(1).reshape((_aux_dips_pos.shape[0], 1)) +
+        #         (_aux_dips_pos ** 2).sum(1).reshape((1, _aux_dips_pos.shape[0])) -
+        #         2 * _aux_dips_pos.dot(_aux_dips_pos.T))).astype("float32")
+        #
+        #     SED_dips_rest = (T.sqrt(
+        #         (_aux_dips_pos ** 2).sum(1).reshape((_aux_dips_pos.shape[0], 1)) +
+        #         (_aux_rest_layer_points ** 2).sum(1).reshape((1, _aux_rest_layer_points.shape[0])) -
+        #         2 * _aux_dips_pos.dot(_aux_rest_layer_points.T))).astype("float32")
+        #
+        #     SED_dips_ref = (T.sqrt(
+        #         (_aux_dips_pos ** 2).sum(1).reshape((_aux_dips_pos.shape[0], 1)) +
+        #         (_aux_ref_layer_points ** 2).sum(1).reshape((1, _aux_ref_layer_points.shape[0])) -
+        #         2 * _aux_dips_pos.dot(_aux_ref_layer_points.T))).astype("float32")
+        #
+        #     # Calculating euclidian distances between the point to simulate and the avalible data
+        #     SED_dips_SimPoint = (T.sqrt(
+        #         (_aux_dips_pos ** 2).sum(1).reshape((_aux_dips_pos.shape[0], 1)) +
+        #         (_aux_grid_val ** 2).sum(1).reshape((1, _aux_grid_val.shape[0])) -
+        #         2 * _aux_dips_pos.dot(_aux_grid_val.T))).astype("float32")
+        #
+        #     SED_rest_SimPoint = (T.sqrt(
+        #         (_aux_rest_layer_points ** 2).sum(1).reshape((_aux_rest_layer_points.shape[0], 1)) +
+        #         (_aux_grid_val ** 2).sum(1).reshape((1, _aux_grid_val.shape[0])) -
+        #         2 * _aux_rest_layer_points.dot(_aux_grid_val.T))).astype("float32")
+        #
+        #     SED_ref_SimPoint = (T.sqrt(
+        #         (_aux_ref_layer_points ** 2).sum(1).reshape((_aux_ref_layer_points.shape[0], 1)) +
+        #         (_aux_grid_val ** 2).sum(1).reshape((1, _aux_grid_val.shape[0])) -
+        #         2 * _aux_ref_layer_points.dot(_aux_grid_val.T))).astype("float32")
+        #
+        #     # Cartesian distances between dips positions
+        #     h_u = T.vertical_stack(
+        #         T.tile(dips_position[:, 0] - dips_position[:, 0].reshape((dips_position[:, 0].shape[0], 1)),
+        #                n_dimensions),
+        #         T.tile(dips_position[:, 1] - dips_position[:, 1].reshape((dips_position[:, 1].shape[0], 1)),
+        #                n_dimensions),
+        #         T.tile(dips_position[:, 2] - dips_position[:, 2].reshape((dips_position[:, 2].shape[0], 1)),
+        #                n_dimensions))
+        #
+        #     h_v = h_u.T
+        #
+        #     # Cartesian distances between dips and interface points
+        #     # Rest
+        #     hu_rest = T.vertical_stack(
+        #         (dips_position[:, 0] - rest_layer_points[:, 0].reshape((rest_layer_points[:, 0].shape[0], 1))).T,
+        #         (dips_position[:, 1] - rest_layer_points[:, 1].reshape((rest_layer_points[:, 1].shape[0], 1))).T,
+        #         (dips_position[:, 2] - rest_layer_points[:, 2].reshape((rest_layer_points[:, 2].shape[0], 1))).T
+        #     )
+        #
+        #     # Reference point
+        #     hu_ref = T.vertical_stack(
+        #         (dips_position[:, 0] - ref_layer_points[:, 0].reshape((ref_layer_points[:, 0].shape[0], 1))).T,
+        #         (dips_position[:, 1] - ref_layer_points[:, 1].reshape((ref_layer_points[:, 1].shape[0], 1))).T,
+        #         (dips_position[:, 2] - ref_layer_points[:, 2].reshape((ref_layer_points[:, 2].shape[0], 1))).T
+        #     )
+        #
+        #     # Cartesian distances between reference points and rest
+        #     hx = T.stack(
+        #         (rest_layer_points[:, 0] - ref_layer_points[:, 0]),
+        #         (rest_layer_points[:, 1] - ref_layer_points[:, 1]),
+        #         (rest_layer_points[:, 2] - ref_layer_points[:, 2])
+        #     ).T
+        #
+        #     # Cartesian distances between the point to simulate and the dips
+        #     hu_SimPoint = T.vertical_stack(
+        #         (dips_position[:, 0] - grid_val[:, 0].reshape((grid_val[:, 0].shape[0], 1))).T,
+        #         (dips_position[:, 1] - grid_val[:, 1].reshape((grid_val[:, 1].shape[0], 1))).T,
+        #         (dips_position[:, 2] - grid_val[:, 2].reshape((grid_val[:, 2].shape[0], 1))).T
+        #     )
+        #
+        #     # Perpendicularity matrix. Boolean matrix to separate cross-covariance and
+        #     # every gradient direction covariance
+        #     perpendicularity_matrix = T.zeros_like(SED_dips_dips)
+        #
+        #     # Cross-covariances of x
+        #     perpendicularity_matrix = T.set_subtensor(
+        #         perpendicularity_matrix[0:dips_position.shape[0], 0:dips_position.shape[0]], 1)
+        #
+        #     # Cross-covariances of y
+        #     perpendicularity_matrix = T.set_subtensor(
+        #         perpendicularity_matrix[dips_position.shape[0]:dips_position.shape[0] * 2,
+        #         dips_position.shape[0]:dips_position.shape[0] * 2], 1)
+        #
+        #     # Cross-covariances of y
+        #     perpendicularity_matrix = T.set_subtensor(
+        #         perpendicularity_matrix[dips_position.shape[0] * 2:dips_position.shape[0] * 3,
+        #         dips_position.shape[0] * 2:dips_position.shape[0] * 3], 1)
+        #
+        #
+        #    # printing = (self.c_o_T * i_reescale * (
+        #    #     (SED_rest_rest < self.a_T) *  # Rest - Rest Covariances Matrix
+        #     #printing = (1 - 7 * (SED_rest_rest / self.a_T) ** 2 +
+        #     #     35 / 4 * (SED_rest_rest / self.a_T) ** 3 -
+        #     #     7 / 2 * (SED_rest_rest / self.a_T) ** 5 +
+        #     #     3 / 4 * (SED_rest_rest / self.a_T) ** 7)
+        #     printing = SED_rest_rest*self.a_T
+        #     # ==========================
+        #     # Creating covariance Matrix
+        #     # ==========================
+        #     # Covariance matrix for interfaces
+        #     C_I = (self.c_o_T * i_reescale * (
+        #         (SED_rest_rest < self.a_T) *  # Rest - Rest Covariances Matrix
+        #         (1 - 7 * (SED_rest_rest / self.a_T) ** 2 +
+        #          35 / 4 * (SED_rest_rest / self.a_T) ** 3 -
+        #          7 / 2 * (SED_rest_rest / self.a_T) ** 5 +
+        #          3 / 4 * (SED_rest_rest / self.a_T) ** 7) -
+        #         ((SED_ref_rest < self.a_T) *  # Reference - Rest
+        #          (1 - 7 * (SED_ref_rest / self.a_T) ** 2 +
+        #           35 / 4 * (SED_ref_rest / self.a_T) ** 3 -
+        #           7 / 2 * (SED_ref_rest / self.a_T) ** 5 +
+        #           3 / 4 * (SED_ref_rest / self.a_T) ** 7)) -
+        #         ((SED_rest_ref < self.a_T) *  # Rest - Reference
+        #          (1 - 7 * (SED_rest_ref / self.a_T) ** 2 +
+        #           35 / 4 * (SED_rest_ref / self.a_T) ** 3 -
+        #           7 / 2 * (SED_rest_ref / self.a_T) ** 5 +
+        #           3 / 4 * (SED_rest_ref / self.a_T) ** 7)) +
+        #         ((SED_ref_ref < self.a_T) *  # Reference - References
+        #          (1 - 7 * (SED_ref_ref / self.a_T) ** 2 +
+        #           35 / 4 * (SED_ref_ref / self.a_T) ** 3 -
+        #           7 / 2 * (SED_ref_ref / self.a_T) ** 5 +
+        #           3 / 4 * (SED_ref_ref / self.a_T) ** 7)))) #'+ 10e-6
+        #
+        #     SED_dips_dips = T.switch(T.eq(SED_dips_dips, 0), 1, SED_dips_dips)
+        #
+        #     # Covariance matrix for gradients at every xyz direction and their cross-covariances
+        #     C_G = T.switch(
+        #         T.eq(SED_dips_dips, 0),  # This is the condition
+        #         0,  # If true it is equal to 0. This is how a direction affect another
+        #         (  # else, following Chiles book
+        #             (h_u * h_v / SED_dips_dips ** 2) *
+        #             ((
+        #                  (SED_dips_dips < self.a_T) *  # first derivative
+        #                  (-self.c_o_T * ((-14 / self.a_T ** 2) + 105 / 4 * SED_dips_dips / self.a_T ** 3 -
+        #                                  35 / 2 * SED_dips_dips ** 3 / self.a_T ** 5 +
+        #                                  21 / 4 * SED_dips_dips ** 5 / self.a_T ** 7))) +
+        #              (SED_dips_dips < self.a_T) *  # Second derivative
+        #              self.c_o_T * 7 * (9 * SED_dips_dips ** 5 - 20 * self.a_T ** 2 * SED_dips_dips ** 3 +
+        #                                15 * self.a_T ** 4 * SED_dips_dips - 4 * self.a_T ** 5) / (2 * self.a_T ** 7)) -
+        #             (perpendicularity_matrix *
+        #              (SED_dips_dips < self.a_T) *  # first derivative
+        #              self.c_o_T * ((-14 / self.a_T ** 2) + 105 / 4 * SED_dips_dips / self.a_T ** 3 -
+        #                            35 / 2 * SED_dips_dips ** 3 / self.a_T ** 5 +
+        #                            21 / 4 * SED_dips_dips ** 5 / self.a_T ** 7)))
+        #     )
+        #
+        #     # Setting nugget effect of the gradients
+        #     # TODO: This function can be substitued by simply adding the nugget effect to the diag
+        #     C_G = T.fill_diagonal(C_G, -self.c_o_T * (-14 / self.a_T ** 2) + self.nugget_effect_grad_T)
+        #
+        #     # Cross-Covariance gradients-interfaces
+        #     C_GI = gi_reescale * (
+        #         (hu_rest *
+        #          (SED_dips_rest < self.a_T) *  # first derivative
+        #          (- self.c_o_T * ((-14 / self.a_T ** 2) + 105 / 4 * SED_dips_rest / self.a_T ** 3 -
+        #                           35 / 2 * SED_dips_rest ** 3 / self.a_T ** 5 +
+        #                           21 / 4 * SED_dips_rest ** 5 / self.a_T ** 7))) -
+        #         (hu_ref *
+        #          (SED_dips_ref < self.a_T) *  # first derivative
+        #          (- self.c_o_T * ((-14 / self.a_T ** 2) + 105 / 4 * SED_dips_ref / self.a_T ** 3 -
+        #                           35 / 2 * SED_dips_ref ** 3 / self.a_T ** 5 +
+        #                           21 / 4 * SED_dips_ref ** 5 / self.a_T ** 7)))
+        #     ).T
+        #
+        #     if self.u_grade_T.get_value() == 3:
+        #         # ==========================
+        #         # Condition of universality 1 degree
+        #
+        #         # Gradients
+        #         n = dips_position.shape[0]
+        #         U_G = T.zeros((n * n_dimensions, n_dimensions))
+        #         # x
+        #         U_G = T.set_subtensor(
+        #             U_G[:n, 0], 1)
+        #         # y
+        #         U_G = T.set_subtensor(
+        #             U_G[n:n * 2, 1], 1
+        #         )
+        #         # z
+        #         U_G = T.set_subtensor(
+        #             U_G[n * 2: n * 3, 2], 1
+        #         )
+        #
+        #     #    U_G = T.set_subtensor(U_G[:, -1], [0, 0, 0, 0, 0, 0])
+        #
+        #         # Interface
+        #         U_I = -hx * gi_reescale
+        #
+        #     #    hxf = (T.lt(rest_layer_points[:, 0], 5) - T.lt(ref_layer_points[:, 0], 5))*2 + 1
+        #
+        #     #    U_I = T.horizontal_stack(U_I, T.stack(hxf).T)
+        #
+        #     elif self.u_grade_T.get_value() == 9:
+        #         # ==========================
+        #         # Condition of universality 2 degree
+        #         # Gradients
+        #
+        #         n = dips_position.shape[0]
+        #         U_G = T.zeros((n * n_dimensions, 3 * n_dimensions))
+        #         # x
+        #         U_G = T.set_subtensor(U_G[:n, 0], 1)
+        #         # y
+        #         U_G = T.set_subtensor(U_G[n * 1:n * 2, 1], 1)
+        #         # z
+        #         U_G = T.set_subtensor(U_G[n * 2: n * 3, 2], 1)
+        #         # x**2
+        #         U_G = T.set_subtensor(U_G[:n, 3], 2 * gi_reescale * dips_position[:, 0])
+        #         # y**2
+        #         U_G = T.set_subtensor(U_G[n * 1:n * 2, 4], 2 * gi_reescale * dips_position[:, 1])
+        #         # z**2
+        #         U_G = T.set_subtensor(U_G[n * 2: n * 3, 5], 2 * gi_reescale * dips_position[:, 2])
+        #         # xy
+        #         U_G = T.set_subtensor(U_G[:n, 6], gi_reescale * dips_position[:, 1])  # This is y
+        #         U_G = T.set_subtensor(U_G[n * 1:n * 2, 6], gi_reescale * dips_position[:, 0])  # This is x
+        #         # xz
+        #         U_G = T.set_subtensor(U_G[:n, 7], gi_reescale * dips_position[:, 2])  # This is z
+        #         U_G = T.set_subtensor(U_G[n * 2: n * 3, 7], gi_reescale * dips_position[:, 0])  # This is x
+        #         # yz
+        #         U_G = T.set_subtensor(U_G[n * 1:n * 2, 8], gi_reescale * dips_position[:, 2])  # This is z
+        #         U_G = T.set_subtensor(U_G[n * 2:n * 3, 8], gi_reescale * dips_position[:, 1])  # This is y
+        #
+        #         # Interface
+        #         U_I = - T.stack(
+        #             gi_reescale * (rest_layer_points[:, 0] - ref_layer_points[:, 0]), # x
+        #             gi_reescale * (rest_layer_points[:, 1] - ref_layer_points[:, 1]), # y
+        #             gi_reescale * (rest_layer_points[:, 2] - ref_layer_points[:, 2]), # z
+        #             gi_reescale ** 2 * (rest_layer_points[:, 0] ** 2 - ref_layer_points[:, 0] ** 2), # xx
+        #             gi_reescale ** 2 * (rest_layer_points[:, 1] ** 2 - ref_layer_points[:, 1] ** 2), # yy
+        #             gi_reescale ** 2 * (rest_layer_points[:, 2] ** 2 - ref_layer_points[:, 2] ** 2), # zz
+        #             gi_reescale ** 2 * (rest_layer_points[:, 0] * rest_layer_points[:, 1] - ref_layer_points[:, 0] *
+        #                                 ref_layer_points[:, 1]),
+        #             gi_reescale ** 2 * (rest_layer_points[:, 0] * rest_layer_points[:, 2] - ref_layer_points[:, 0] *
+        #                                 ref_layer_points[:, 2]),
+        #             gi_reescale ** 2 * (rest_layer_points[:, 1] * rest_layer_points[:, 2] - ref_layer_points[:, 1] *
+        #                                 ref_layer_points[:, 2]),
+        #         ).T
+        #
+        #
+        #
+        #     # =================================
+        #     # Creation of the Covariance Matrix
+        #     # =================================
+        #     C_matrix = T.zeros((length_of_C, length_of_C ))
+        #
+        #     # First row of matrices
+        #     C_matrix = T.set_subtensor(C_matrix[0:length_of_CG, 0:length_of_CG], C_G)
+        #
+        #     C_matrix = T.set_subtensor(C_matrix[0:length_of_CG, length_of_CG:length_of_CG + length_of_CGI], C_GI.T)
+        #
+        #     if not self.u_grade_T.get_value() == 0:
+        #         C_matrix = T.set_subtensor(C_matrix[0:length_of_CG, -length_of_U_I:], U_G)
+        #
+        #     # Second row of matrices
+        #     C_matrix = T.set_subtensor(C_matrix[length_of_CG:length_of_CG + length_of_CGI, 0:length_of_CG], C_GI)
+        #     C_matrix = T.set_subtensor(C_matrix[length_of_CG:length_of_CG + length_of_CGI,
+        #                                length_of_CG:length_of_CG + length_of_CGI], C_I)
+        #
+        #     if not self.u_grade_T.get_value() == 0:
+        #         C_matrix = T.set_subtensor(C_matrix[length_of_CG:length_of_CG + length_of_CGI, -length_of_U_I:], U_I)
+        #
+        #         # Third row of matrices
+        #         C_matrix = T.set_subtensor(C_matrix[-length_of_U_I:, 0:length_of_CG], U_G.T)
+        #         C_matrix = T.set_subtensor(C_matrix[-length_of_U_I:, length_of_CG:length_of_CG + length_of_CGI], U_I.T)
+        #
+        #     # =====================
+        #     # Creation of the gradients G vector
+        #     # Calculation of the cartesian components of the dips assuming the unit module
+        #     G_x = T.sin(T.deg2rad(dip_angles)) * T.sin(T.deg2rad(azimuth)) * polarity
+        #     G_y = T.sin(T.deg2rad(dip_angles)) * T.cos(T.deg2rad(azimuth)) * polarity
+        #     G_z = T.cos(T.deg2rad(dip_angles)) * polarity
+        #
+        #     G = T.concatenate((G_x, G_y, G_z))
+        #
+        #     # Creation of the Dual Kriging vector
+        #     b = T.zeros_like(C_matrix[:, 0])
+        #     b = T.set_subtensor(b[0:G.shape[0]], G)
+        #
+        #     # Solving the kriging system
+        #     # TODO: look for an eficient way to substitute nlianlg by a theano operation
+        #     DK_parameters = T.dot(T.nlinalg.matrix_inverse(C_matrix), b)
+        #
+        #     # ==============
+        #     # Interpolator
+        #     # ==============
+        #
+        #     # Creation of a matrix of dimensions equal to the grid with the weights for every point (big 4D matrix in
+        #     # ravel form)
+        #     weights = T.tile(DK_parameters, (grid_val.shape[0], 1)).T
+        #
+        #     # Gradient contribution
+        #     sigma_0_grad = T.sum(
+        #         (weights[:length_of_CG, :] *
+        #          gi_reescale *
+        #          (-hu_SimPoint *
+        #           (SED_dips_SimPoint < self.a_T) *  # first derivative
+        #           (- self.c_o_T * ((-14 / self.a_T ** 2) + 105 / 4 * SED_dips_SimPoint / self.a_T ** 3 -
+        #                            35 / 2 * SED_dips_SimPoint ** 3 / self.a_T ** 5 +
+        #                            21 / 4 * SED_dips_SimPoint ** 5 / self.a_T ** 7)))),
+        #         axis=0)
+        #
+        #     # Interface contribution
+        #     sigma_0_interf = (T.sum(
+        #         -weights[length_of_CG:length_of_CG + length_of_CGI, :] *
+        #         (self.c_o_T * i_reescale * (
+        #             (SED_rest_SimPoint < self.a_T) *  # SimPoint - Rest Covariances Matrix
+        #             (1 - 7 * (SED_rest_SimPoint / self.a_T) ** 2 +
+        #              35 / 4 * (SED_rest_SimPoint / self.a_T) ** 3 -
+        #              7 / 2 * (SED_rest_SimPoint / self.a_T) ** 5 +
+        #              3 / 4 * (SED_rest_SimPoint / self.a_T) ** 7) -
+        #             ((SED_ref_SimPoint < self.a_T) *  # SimPoint- Ref
+        #              (1 - 7 * (SED_ref_SimPoint / self.a_T) ** 2 +
+        #               35 / 4 * (SED_ref_SimPoint / self.a_T) ** 3 -
+        #               7 / 2 * (SED_ref_SimPoint / self.a_T) ** 5 +
+        #               3 / 4 * (SED_ref_SimPoint / self.a_T) ** 7)))), axis=0))
+        #
+        #     # Universal drift contribution
+        #     # Universal terms used to calculate f0
+        #     _universal_terms_layers = T.horizontal_stack(
+        #         rest_layer_points,
+        #         (rest_layer_points ** 2),
+        #         T.stack((rest_layer_points[:, 0] * rest_layer_points[:, 1],
+        #                  rest_layer_points[:, 0] * rest_layer_points[:, 2],
+        #                  rest_layer_points[:, 1] * rest_layer_points[:, 2]), axis=1)).T
+        #
+        #     universal_matrix = T.horizontal_stack(
+        #         (self.universal_matrix_T * yet_simulated).nonzero_values().reshape((9, -1)),
+        #         _universal_terms_layers)
+        #
+        #     if self.u_grade_T.get_value() == 0:
+        #         f_0 = 0
+        #     else:
+        #         gi_rescale_aux = T.repeat(gi_reescale, 9)
+        #         gi_rescale_aux = T.set_subtensor(gi_rescale_aux[:3], 1)
+        #         _aux_magic_term = T.tile(gi_rescale_aux[:grade_universal], (grid_val.shape[0], 1)).T
+        #         f_0 = (T.sum(
+        #             weights[-length_of_U_I:, :] * gi_reescale * _aux_magic_term *
+        #             universal_matrix[:grade_universal]
+        #             , axis=0))
+        #
+        #     # Contribution faults
+        #    # f_1 = weights[-1, :] * T.lt(universal_matrix[0, :], 5) * 2 - 1
+        #  #   f_1 = 0
+        #     # Potential field
+        #     # Value of the potential field
+        #
+        #     Z_x = (sigma_0_grad + sigma_0_interf + f_0)[:-rest_layer_points.shape[0]]
+        #     potential_field_interfaces = (sigma_0_grad + sigma_0_interf + f_0)[-rest_layer_points.shape[0]:]
+        #
+        #     # Theano function to calculate a potential field
+        #     self._interpolate = theano.function(
+        #         [dips_position, dip_angles, azimuth, polarity, rest_layer_points, ref_layer_points,
+        #          theano.In(yet_simulated, value=np.ones_like(self._grid.grid[:, 0]))],
+        #         [Z_x, G_x, G_y, G_z, potential_field_interfaces, C_matrix, printing],
+        #         on_unused_input="warn", profile=True, allow_input_downcast=True)
+        #
+        #     # =======================================================================
+        #     #               CODE TO EXPORT THE BLOCK DIRECTLY
+        #     # ========================================================================
+        #
+        #     # Aux shared parameters
+        #    # infinite_pos = theano.shared(np.float32(np.inf))
+        #    # infinite_neg = theano.shared(np.float32(-np.inf))
+        #
+        #     # Value of the lithology-segment
+        #     n_formation = T.vector("The assigned number of the lithologies in this serie")
+        #
+        #     # Loop to obtain the average Zx for every intertace
+        #     def average_potential(dim_a, dim_b, pfi):
+        #         """
+        #
+        #         :param dim: size of the rest values vector per formation
+        #         :param pfi: the values of all the rest values potentials
+        #         :return: average of the potential per formation
+        #         """
+        #         average = pfi[T.cast(dim_a, "int32"): T.cast(dim_b, "int32")].sum() / (dim_b - dim_a)
+        #         return average
+        #
+        #     potential_field_unique, updates1 = theano.scan(fn=average_potential,
+        #                                                    outputs_info=None,
+        #                                                    sequences=dict(
+        #                                                        input=T.concatenate(
+        #                                                            (T.stack(0),
+        #                                                             self.number_of_points_per_formation_T)),
+        #                                                        taps=[0, 1]),
+        #                                                    non_sequences=potential_field_interfaces)
+        #
+        #     infinite_pos = T.max(potential_field_unique) + 10
+        #     infinite_neg = T.min(potential_field_unique) - 10
+        #
+        #     # Loop to segment the distinct lithologies
+        #     potential_field_iter = T.concatenate((T.stack(infinite_pos),
+        #                                           potential_field_unique,
+        #                                           T.stack(infinite_neg)))
+        #
+        #     def compare(a, b, n_formation, Zx):
+        #         return T.le(Zx, a) * T.ge(Zx, b) * n_formation
+        #
+        #     block, updates2 = theano.scan(fn=compare,
+        #                                   outputs_info=None,
+        #                                   sequences=[dict(input=potential_field_iter, taps=[0, 1]),
+        #                                              n_formation],
+        #                                   non_sequences=Z_x)
+        #
+        #     # Adding to the block the contribution of the potential field
+        #     potential_field_contribution = T.set_subtensor(
+        #         self.block[T.nonzero(T.cast(yet_simulated, "int8"))[0]],
+        #         block.sum(axis=0))
+        #
+        #     # Some gradient testing
+        #     # grad = T.jacobian(T.flatten(printing), rest_layer_points)
+        #     # grad = T.grad(T.sum(Z_x), self.a_T)
+        #     from theano.compile.nanguardmode import NanGuardMode
+        #
+        #     def detect_nan(i, node, fn):
+        #         for output in fn.outputs:
+        #             if (not isinstance(output[0], np.random.RandomState) and
+        #                     np.isnan(output[0]).any()):
+        #                 print('*** NaN detected ***')
+        #                 theano.printing.debugprint(node)
+        #                 print('Inputs : %s' % [input[0] for input in fn.inputs])
+        #                 print('Outputs: %s' % [output[0] for output in fn.outputs])
+        #                 break
+        #
+        #     # Theano function to update the block
+        #     self._block_export = theano.function([dips_position, dip_angles, azimuth, polarity, rest_layer_points,
+        #                                           ref_layer_points, n_formation, yet_simulated], None,
+        #                                          updates=[(self.block, potential_field_contribution)],
+        #                                          on_unused_input="warn", profile=True, allow_input_downcast=True,)
+        #                                      #   mode=theano.compile.MonitorMode(
+        #                                      #       post_func=detect_nan))
+        #                                       #    mode=NanGuardMode(nan_is_error=True, inf_is_error=True,
+        #                                       #                     big_is_error=True))
